fastapi/requirements.txt
redis==4.5.4
elasticsearch[async]==7.9.1
fastapi==0.92.0
orjson==3.9.10
pydantic==1.10.5
python-dotenv==1.0.0
pre-commit==3.6.0
uvicorn==0.12.2
fastapi-cache2[redis]==0.2.1
#docker
uvloop==0.17.0
#local(windows)
#uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
